requirements.txt
easyocr==1.7.1
Flask==3.0.3
Flask-Cors==4.0.1
numpy==1.26.4
Pyrebase4==4.8.0
python-dotenv==1.0.1
pymongo